[#7] refactor: fix import paths
