+++ src/pages/admin/matching/tabs/SupporterTab.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import SupporterStatusTable from "../../../../components/admin/matching/table/SupporterStatusTable";
+import SupporterStatusTable from "../../../../components/admin/matching/Supporter/SupporterStatusTable";
 import { Supporter } from "../../../../types/user";
 
 interface SupporterTabProps {
